requirements.txt
aiodns==3.0.0
aiohttp==3.8.4
aiohttp-json-rpc==0.13.3
aiosignal==1.3.1
aiosqlite==0.17.0
appdirs==1.4.4
apsw==3.41.2.0
apsw-wheels==3.36.0.post1
async-timeout==4.0.2
attrs==23.1.0
Babel==2.12.1
Brotli==1.0.9
certifi==2022.12.7
cffi==1.15.1
chardet==4.0.0
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.4
commonmark==0.9.1
contextlib2==21.6.0
discord.py==2.2.3
distro==1.8.0
frozenlist==1.3.3
fuzzywuzzy==0.18.0
idna==3.4
importlib-metadata==6.6.0
Markdown==3.4.3
markdown-it-py==2.2.0
mdurl==0.1.2
multidict==6.0.4
openai==0.27.0
orjson==3.8.11
packaging==23.1
platformdirs==3.5.0
psutil==5.9.5
pycares==4.3.0
pycparser==2.21
Pygments==2.15.1
PyNaCl==1.4.0
python-dateutil==2.8.2
python-Levenshtein-wheels==0.13.2
pytz==2023.3
PyYAML==6.0
rapidfuzz==3.0.0
red-commons==1.0.0
Red-DiscordBot==3.5.1
Red-Lavalink==0.11.0
requests==2.28.2
rich==13.3.5
schema==0.7.5
six==1.16.0
tqdm==4.65.0
typing-extensions==4.5.0
urllib3==1.26.14
uvloop==0.17.0
yarl==1.9.2
zipp==3.15.0
